+++ spec/bounding.spec.js
@@ -12,7 +12,7 @@ describe('BoundToRectangle', () => {
   it('should bound point to rectangle', () => {
     const bounding = new BoundToRectangle(new Rectangle(new Point(30, 30), new Point(100, 100)))
     let point = new Point(0, 0)
-    let size = new Point(25, 25)
+    const size = new Point(25, 25)
 
     expect(bounding.bound(point, size).x).toEqual(30)
     expect(bounding.bound(point, size).y).toEqual(30)
@@ -27,7 +27,7 @@ describe('BoundToLineX', () => {
   it('should bound point to line', () => {
     const bounding = new BoundToLineX(55, 100, 500)
     let point = new Point(0, 0)
-    let size = new Point(25, 25)
+    const size = new Point(25, 25)
 
     expect(bounding.bound(point, size).x).toEqual(55)
     expect(bounding.bound(point, size).y).toEqual(100)
@@ -46,7 +46,7 @@ describe('BoundToLineY', () => {
   it('should bound point to line', () => {
     const bounding = new BoundToLineY(55, 100, 500)
     let point = new Point(0, 0)
-    let size = new Point(25, 25)
+    const size = new Point(25, 25)
 
     expect(bounding.bound(point, size).x).toEqual(100)
     expect(bounding.bound(point, size).y).toEqual(55)
@@ -65,7 +65,7 @@ describe('BoundToLine', () => {
   it('should bound point to line', () => {
     const bounding = new BoundToLine(new Point(0, 0), new Point(200, 200))
     let point = new Point(0, 0)
-    let size = new Point(25, 25)
+    const size = new Point(25, 25)
 
     expect(bounding.bound(point, size).x).toEqual(0)
     expect(bounding.bound(point, size).y).toEqual(0)
@@ -84,7 +84,7 @@ describe('BoundToCircle', () => {
   it('should bound point on circle', () => {
     const bounding = new BoundToCircle(new Point(100, 100), 70)
     let point = new Point(100, 0)
-    let size = new Point(25, 25)
+    const size = new Point(25, 25)
 
     expect(bounding.bound(point, size).x).toEqual(100)
     expect(bounding.bound(point, size).y).toEqual(30)